Make definitions in getMembers() configurable
Moved the definitions out of getMembers() to make it configurable.  It allows more flexibility to template developers and plugin developers to create their categories of data as needed.

Functions
addMemberDefinition(): Allows people to add new categories/kind to the data and pull that data as a top level item for navigation. 

renameMemberDefinition(): Allows for relabeling existing categories just in case someone wants/needs to.